move Navbar to containers dir
