+++ client/src/Root.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom';
 
 // Components
-import Navbar from './components/presentation/Navbar';
+import Navbar from './components/containers/Navbar';
 import Landing from './components/presentation/Landing';
 import Register from './components/containers/Register';
 import Login from './components/containers/Login';
